use static lib to reduce build time; move completion code to cli
